AITutorAgent-asmitha/requirements.txt
streamlit
google-generativeai
google-api-python-client
wikipedia
youtube-search-python
langchain
python-dotenv
google-generativeai
google-api-python-client
wikipedia
youtube-search-python